Changed Cost of living reads from int to float

+++ Your_Location_In_The_Nation/src/FakeDatabase/InitialData.java
@@ -33,9 +33,9 @@ public class InitialData {
 				Loc.setState(i.next());
 				Loc.setZipcode(i.next());
 				Loc.setAvgSalaryPerHouse(Integer.parseInt(i.next()));
-				Loc.setCostOfLivingRent(Integer.parseInt(i.next()));
-				Loc.setCostOfLivingOwnWithMortgage(Integer.parseInt(i.next()));
-				Loc.setCostOfLivingOwnNoMortgage(Integer.parseInt(i.next()));
+				Loc.setCostOfLivingRent(Float.parseFloat(i.next()));
+				Loc.setCostOfLivingOwnWithMortgage(Float.parseFloat(i.next()));
+				Loc.setCostOfLivingOwnNoMortgage(Float.parseFloat(i.next()));
 				Loc.setCrimeRate(Integer.parseInt(i.next()));
 				Loc.setRegion(i.next());
 				Loc.setPopulation(i.next());
